requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-cors-headers==4.3.0
python-dotenv==1.0.0
Pillow==10.0.0
django-filter==23.2 